Move rule_based under representation directory
