+++ leaf_common/representation/rule_based/condition.py
@@ -6,7 +6,7 @@ from typing import Dict
 from typing import List
 from typing import Tuple
 
-from leaf_common.rule_based.rules_evaluation_constants \
+from leaf_common.representation.rule_based.rules_evaluation_constants \
     import RulesEvaluationConstants
