+++ leaf_common/representation/rule_based/rules_evaluation_constants.py
