+++ tests/fixtures/saved_agent.rules
@@ -9010,7 +9010,7 @@
     "0": 0,
     "1": 1
   },
-  "py/object": "leaf_common.rule_based.rules_agent.RulesAgent",
+  "py/object": "leaf_common.representation.rule_based.rules_agent.RulesAgent",
   "rules": [
     {
       "action": "0",
@@ -9025,7 +9025,7 @@
           "first_state_coefficient": 0.98,
           "first_state_lookback": 0,
           "operator": ">",
-          "py/object": "leaf_common.rule_based.condition.Condition",
+          "py/object": "leaf_common.representation.rule_based.condition.Condition",
           "second_state_key": "value",
           "second_state_coefficient": 0.72,
           "second_state_lookback": 0,
@@ -9046,7 +9046,7 @@
           "first_state_coefficient": 0.69,
           "first_state_lookback": 0,
           "operator": "<=",
-          "py/object": "leaf_common.rule_based.condition.Condition",
+          "py/object": "leaf_common.representation.rule_based.condition.Condition",
           "second_state_key": "0",
           "second_state_coefficient": 0.46,
           "second_state_lookback": 0,
@@ -9067,7 +9067,7 @@
           "first_state_coefficient": 0.48,
           "first_state_lookback": 0,
           "operator": ">=",
-          "py/object": "leaf_common.rule_based.condition.Condition",
+          "py/object": "leaf_common.representation.rule_based.condition.Condition",
           "second_state_key": "1",
           "second_state_coefficient": 0.08,
           "second_state_lookback": 0,
@@ -9088,7 +9088,7 @@
           "first_state_coefficient": 0.26,
           "first_state_lookback": 0,
           "operator": "<",
-          "py/object": "leaf_common.rule_based.condition.Condition",
+          "py/object": "leaf_common.representation.rule_based.condition.Condition",
           "second_state_key": "2",
           "second_state_coefficient": 0.63,
           "second_state_lookback": 0,
@@ -9105,7 +9105,7 @@
           }
         }
       ],
-      "py/object": "leaf_common.rule_based.rule.Rule",
+      "py/object": "leaf_common.representation.rule_based.rule.Rule",
       "times_applied": 736
     },
     {
@@ -9121,7 +9121,7 @@
           "first_state_coefficient": 0.98,
           "first_state_lookback": 0,
           "operator": ">",
-          "py/object": "leaf_common.rule_based.condition.Condition",
+          "py/object": "leaf_common.representation.rule_based.condition.Condition",
           "second_state_key": "value",
           "second_state_coefficient": 0.72,
           "second_state_lookback": 0,
@@ -9142,7 +9142,7 @@
           "first_state_coefficient": 0.69,
           "first_state_lookback": 0,
           "operator": "<=",
-          "py/object": "leaf_common.rule_based.condition.Condition",
+          "py/object": "leaf_common.representation.rule_based.condition.Condition",
           "second_state_key": "0",
           "second_state_coefficient": 0.46,
           "second_state_lookback": 0,
@@ -9163,7 +9163,7 @@
           "first_state_coefficient": 0.48,
           "first_state_lookback": 0,
           "operator": ">=",
-          "py/object": "leaf_common.rule_based.condition.Condition",
+          "py/object": "leaf_common.representation.rule_based.condition.Condition",
           "second_state_key": "1",
           "second_state_coefficient": 0.08,
           "second_state_lookback": 0,
@@ -9184,7 +9184,7 @@
           "first_state_coefficient": 0.42,
           "first_state_lookback": 0,
           "operator": "<",
-          "py/object": "leaf_common.rule_based.condition.Condition",
+          "py/object": "leaf_common.representation.rule_based.condition.Condition",
           "second_state_key": "5",
           "second_state_coefficient": 0.49,
           "second_state_lookback": 0,
@@ -9205,7 +9205,7 @@
           "first_state_coefficient": 0.37,
           "first_state_lookback": 0,
           "operator": ">=",
-          "py/object": "leaf_common.rule_based.condition.Condition",
+          "py/object": "leaf_common.representation.rule_based.condition.Condition",
           "second_state_key": "1",
           "second_state_coefficient": 0.08,
           "second_state_lookback": 0,
@@ -9226,7 +9226,7 @@
           "first_state_coefficient": 0.19,
           "first_state_lookback": 0,
           "operator": ">=",
-          "py/object": "leaf_common.rule_based.condition.Condition",
+          "py/object": "leaf_common.representation.rule_based.condition.Condition",
           "second_state_key": "0",
           "second_state_coefficient": 0.87,
           "second_state_lookback": 0,
@@ -9243,7 +9243,7 @@
           }
         }
       ],
-      "py/object": "leaf_common.rule_based.rule.Rule",
+      "py/object": "leaf_common.representation.rule_based.rule.Rule",
       "times_applied": 0
     }
   ],
